+++ aquabsd.alps/aquabsd.alps.ogl/main.c
@@ -16,7 +16,7 @@ int load(void) {
 
 	ftime_device = kos_query_device(0, (uint64_t) "aquabsd.alps.ftime");
 
-	if (ftime_device != -1) {
+	if (ftime_device != (uint64_t) -1) {
 		aquabsd_alps_ftime_create = kos_load_device_function(ftime_device, "create");
 		aquabsd_alps_ftime_draw   = kos_load_device_function(ftime_device, "draw"  );
 		aquabsd_alps_ftime_swap   = kos_load_device_function(ftime_device, "swap"  );
@@ -28,7 +28,7 @@ int load(void) {
 
     win_device = kos_query_device(0, (uint64_t) "aquabsd.alps.win");
 
-	if (win_device != -1) {
+	if (win_device != (uint64_t) -1) {
     	aquabsd_alps_win_register_dev_cb = kos_load_device_function(win_device, "register_dev_cb");
 		aquabsd_alps_win_get_draw_win = kos_load_device_function(win_device, "get_draw_win");
 	}
@@ -44,7 +44,7 @@ uint64_t send(uint16_t _cmd, void* data) {
 		context_type_t type = args[0];
 		context_t* context = NULL;
 
-		if (type == CONTEXT_TYPE_WIN && win_device != -1) {
+		if (type == CONTEXT_TYPE_WIN && win_device != (uint64_t) -1) {
 			aquabsd_alps_win_t* win = (void*) args[1];
 			context = create_win_context(win);
 		}
